Check species names

check_names checks if the species names are correct and searches for suggestions if the name is misspelled or not found in the Brazilian Flora 2020 database.

Usage

check_names(data, species, max_distance = 0.1, Kingdom = "Plantae")

Arguments

data

(data.frame) the data.frame imported with the load_florabr function.

species

(character) names of the species to be checked.

max_distance

(numeric) Maximum distance (as a fraction) allowed for searching suggestions when the name is misspelled. It can be any value between 0 and 1. The higher the value, the more suggestions are returned. For more details, see agrep. Default = 0.1.

Kingdom

(character) the kingdom to which the species belong. It can be "Plantae" or "Fungi". Default = "Plantae".

Value

a data.frame with the following columns:

input_name: the species names informed in species argument

Spelling: indicates if the species name is Correct (a perfect match with a species name in the Brazilian Flora 2020), Probably_incorrect (partial match), or Incorrect (no match with any species).

Suggested name: If Spelling is Correct, it is the same as the input_name. If Spelling is Probably_correct, one or more suggested names are listed, found according to the maximum distance. If Spelling is Incorrect, the value is NA.

Distance: The integer Levenshtein edit distance. It represents the number of single-character edits (insertions, deletions, or substitutions) required to transform the input_name into the Suggested_name.

taxonomicStatus: the taxonomic status of the species name ("Accepted" or "Synonym").

nomenclaturalStatus: the nomenclatural status of the species name. This information is not available for all species.

acceptedName: If the species name is not accepted or incorrect, the accepted name of the specie. If the species name is accepted and correct, the same as input_name and Suggested_name.

family: the family of the specie.

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data", package = "florabr")
spp <- c("Butia cattarinensis", "Araucaria angustifolia")
check_names(data = bf_data, species = spp)
#>               input_name           Spelling         Suggested_name Distance
#> 1 Araucaria angustifolia            Correct Araucaria angustifolia        0
#> 2    Butia cattarinensis Probably_incorrect     Butia catarinensis        1
#>   taxonomicStatus nomenclaturalStatus           acceptedName        family
#> 1        Accepted             Correct Araucaria angustifolia Araucariaceae
#> 2        Accepted                <NA>     Butia catarinensis     Arecaceae