Get available attributes to filter species

This function displays all the options available to filter species by its characteristics

get_attributes(data, attribute, kingdom = "Plantae")

Arguments

data

(data.frame) a data.frame imported with the load_florabr function or a data.frame generated with the select_species function.

attribute

(character) the type of characteristic. See detail to see the options.

kingdom

(character) the kingdom to which the species belong. It can be "Plantae" or "Fungi". Default = "Plantae".

Value

a data.frame with the available options to use in the select_species function.

Details

The attribute argument accepts the following options: group, subgroup, family, lifeform, habitat, vegetation, origin, endemism, biome, states, taxonomicstatus or nomenclaturalstatus. These options represent different characteristics of species that can be used for filtering.

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Brazilian Flora data
# Get available biomes to filter species
get_attributes(data = bf_data, kingdom = "Plantae", attribute = "Biome")
#>                 biome
#> 1              Amazon
#> 2     Atlantic_Forest
#> 3            Caatinga
#> 4             Cerrado
#> 5 Not_found_in_brazil
#> 6               Pampa
#> 7            Pantanal
# Get available life forms to filter species
get_attributes(data = bf_data, kingdom = "Plantae", attribute = "lifeForm")
#>               lifeForm
#> 1      Aquatic-Benthos
#> 2      Aquatic-Neuston
#> 3     Aquatic-Plankton
#> 4               Bamboo
#> 5              Cushion
#> 6             Dendroid
#> 7           Dracaenoid
#> 8           Flabellate
#> 9              Foliose
#> 10                Herb
#> 11 Liana/scandent/vine
#> 12                 Mat
#> 13 Not_found_in_brazil
#> 14           Palm_tree
#> 15             Pendent
#> 16               Shrub
#> 17            Subshrub
#> 18           Succulent
#> 19            Thallose
#> 20                Tree
#> 21                Tuft
#> 22             Unknown
#> 23                Weft
# Get available states to filter species
get_attributes(data = bf_data, kingdom = "Plantae", attribute = "States")
#>                 states
#> 1                   AC
#> 2                   AL
#> 3                   AM
#> 4                   AP
#> 5                   BA
#> 6                   CE
#> 7                   DF
#> 8                   ES
#> 9                   GO
#> 10                  MA
#> 11                  MG
#> 12                  MS
#> 13                  MT
#> 14 Not_found_in_brazil
#> 15                  PA
#> 16                  PB
#> 17                  PE
#> 18                  PI
#> 19                  PR
#> 20                  RJ
#> 21                  RN
#> 22                  RO
#> 23                  RR
#> 24                  RS
#> 25                  SC
#> 26                  SE
#> 27                  SP
#> 28                  TO
#> 29             Unknown