Download the latest version of Brazilian Flora 2020 database

This function downloads the latest or an older version of Brazilian Flora 2020 database, merges the information into a single data.frame, and saves this data.frame in the specified directory.

Usage

get_florabr(output_dir, data_version = "latest", overwrite = TRUE)

Arguments

output_dir

(character) a directory to save the data downloaded from Brazilian Flora 2020

data_version

(character) Version of the Brazilian Flora database to download. Use "latest" to get the most recent version, updated weekly. Alternatively, specify an older version (e.g., data_version = "393.319"). Default value is "latest".

overwrite

(logical) If TRUE, data is overwritten. Default = TRUE.

Value

The function downloads the latest version of the Brazilian Flora 2020 database from the official source. It then merges the information into a single data.frame, containing details on species, taxonomy, occurrence, and other relevant data. The merged data.frame is then saved as a file in the specified output directory. The data is saved in a format that allows easy loading using the load_florabr function for further analysis in R.

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

if (FALSE) {
dir.create("brazilianflora") #Create a directory to save data
my_dir <- "brazilianflora" #Set directory to save data
#Download, merge and save data
get_florabr(output_dir = my_dir, data_version = "latest", overwrite = TRUE)
}