Retrieve synonyms for species

Retrieve synonyms for species

get_synonym(data, species)

Arguments

data

(data.frame) the data.frame imported with the load_florabr function

species

(character) names of the species

Value

A data.frame containing unique synonyms of the specified species along with relevant information on taxonomic and nomenclatural statuses.

Examples

data("bf_data") #Load Brazilian Flora data
#Species to extract synonyms
spp <- c("Araucaria angustifolia", "Adesmia paranensis")
spp_synonyms <- get_synonym(data = bf_data, species = spp)
spp_synonyms
#>                       synonym           acceptedName taxonomicStatus
#> 35323    Araucaria brasiliana Araucaria angustifolia         Synonym
#> 35325  Araucaria brasiliensis Araucaria angustifolia         Synonym
#> 35327        Araucaria dioica Araucaria angustifolia         Synonym
#> 35328       Araucaria elegans Araucaria angustifolia         Synonym
#> 35329    Araucaria ridolfiana Araucaria angustifolia         Synonym
#> 35330       Araucaria saviana Araucaria angustifolia         Synonym
#> 35332   Columbea angustifolia Araucaria angustifolia         Synonym
#> 35333     Columbea brasiliana Araucaria angustifolia         Synonym
#> 60644            Pinus dioica Araucaria angustifolia         Synonym
#> 141020     Araucaria bibbiani Araucaria angustifolia         Synonym
#> 141021   Araucaria lindleyana Araucaria angustifolia         Synonym
#> 141041   Araucaria brasiliana Araucaria angustifolia         Synonym
#> 85308    Adesmia psoraleoides     Adesmia paranensis         Synonym
#>             nomenclaturalStatus
#> 35323                      <NA>
#> 35325                      <NA>
#> 35327                      <NA>
#> 35328                      <NA>
#> 35329                      <NA>
#> 35330                      <NA>
#> 35332  Legitimate_but_incorrect
#> 35333  Legitimate_but_incorrect
#> 60644                      <NA>
#> 141020                     <NA>
#> 141021                     <NA>
#> 141041             Illegitimate
#> 85308  Legitimate_but_incorrect