Search for taxa using vernacular names

Search for taxa using vernacular names

Usage

select_by_vernacular(data, names, exact = FALSE)

Arguments

data

(data.frame) the data.frame imported with the load_florabr function or geneterate with the function select_species.

names

(character) vernacular name ("Nome comum") of the species to be searched

exact

(logic) if TRUE, the function will search only for exact matches. For example, if names = "pinheiro" and exact = TRUE, the function will return only the species popularly known as "pinheiro". On the other hand, if names = "pinheiro" and exact = FALSE, the function will return other results as "pinheiro-do-parana". Default = FALSE

Value

a data.frame with the species with vernacular names that match the input names

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Brazilian Flora data
#Search for species whose vernacular name is 'pinheiro'
pinheiro_exact <- select_by_vernacular(data = bf_data,
                                       names = "pinheiro",
                                       exact = TRUE)
pinheiro_exact
#>                     species
#> 134008 Araucaria columnaris
#> 140668    Pinus canariensis
#> 140670      Pinus banksiana
#> 140684      Pinus palustris
#> 140692          Pinus nigra
#> 140718       Pinus echinata
#> 140730     Pinus densiflora
#> 140814    Pinus tecunumanii
#> 140819     Pinus thunbergii
#>                                               scientificName acceptedName
#> 134008               Araucaria columnaris (J.R.Forst.) Hook.         <NA>
#> 140668                               Pinus canariensis C.Sm.         <NA>
#> 140670                                 Pinus banksiana Lamb.         <NA>
#> 140684                                 Pinus palustris Mill.         <NA>
#> 140692                                Pinus nigra J.F.Arnold         <NA>
#> 140718                                  Pinus echinata Mill.         <NA>
#> 140730                      Pinus densiflora Siebold & Zucc.         <NA>
#> 140814 Pinus tecunumanii F.Schwerdtf. ex Eguiluz & J.P.Perry         <NA>
#> 140819                                Pinus thunbergii Parl.         <NA>
#>        kingdom       Group Subgroup        family     genus lifeForm
#> 134008 Plantae Gymnosperms     <NA> Araucariaceae Araucaria     Tree
#> 140668 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140670 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140684 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140692 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140718 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140730 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140814 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#> 140819 Plantae Gymnosperms     <NA>      Pinaceae     Pinus     Tree
#>            habitat Biome                  States vegetationType      Origin
#> 134008 Terrestrial       MG;PB;PE;PR;RS;SC;SE;SP Anthropic_Area  Cultivated
#> 140668 Terrestrial                   MG;RJ;RS;SP Anthropic_Area  Cultivated
#> 140670 Terrestrial                            RS Anthropic_Area  Cultivated
#> 140684 Terrestrial                   PR;RS;SC;SP Anthropic_Area  Cultivated
#> 140692 Terrestrial                         RS;SP Anthropic_Area  Cultivated
#> 140718 Terrestrial                      PR;RS;SP Anthropic_Area  Cultivated
#> 140730 Terrestrial                            MG Anthropic_Area  Cultivated
#> 140814 Terrestrial                            SP Anthropic_Area Naturalized
#> 140819 Terrestrial                         PR;SP Anthropic_Area  Cultivated
#>           Endemism taxonomicStatus nomenclaturalStatus
#> 134008 Non-endemic        Accepted             Correct
#> 140668 Non-endemic        Accepted                <NA>
#> 140670 Non-endemic        Accepted                <NA>
#> 140684 Non-endemic        Accepted                <NA>
#> 140692 Non-endemic        Accepted                <NA>
#> 140718 Non-endemic        Accepted                <NA>
#> 140730 Non-endemic        Accepted                <NA>
#> 140814 Non-endemic        Accepted                <NA>
#> 140819 Non-endemic        Accepted                <NA>
#>                                                                                                                                                     vernacularName
#> 134008                                                        pinheiro, pinheiro-de-natal, pinheiro-alemao, pinheiro-de-cook, araucaria-colunar, araucaria-excelsa
#> 140668                                                                                                                             pinheiro-das-canarias, pinheiro
#> 140670                                                                                                                                  pinus, pinheiro, jack-pine
#> 140684                          pinheiro, pinheiro-de-folha-longa, pinheiro-da-florida, pinheiro-amarelo, pinheiro-breu, pinheiro-da-georgia, pinheiro-terebintina
#> 140692                                                                                                                                     pinheiro, austrian-pine
#> 140718 pinheiro-de-folha-curta, pinheiro-da-carolina-do-norte, pinheiro-pobre, pinheiro-amarelo-de-folha-curta, pinheiro-amarelo, pinus, pinheiro, short-leaf-pine
#> 140730                                                                                                                                 pinheiro, japanese-red-pine
#> 140814                                                                                                                                             pinus, pinheiro
#> 140819                                                                                                                                             pinus, pinheiro
#>        taxonRank
#> 134008   Species
#> 140668   Species
#> 140670   Species
#> 140684   Species
#> 140692   Species
#> 140718   Species
#> 140730   Species
#> 140814   Species
#> 140819   Species
#Search for species whose vernacular name is 'pinheiro', allowing non-exact
#matches
pinheiro_not_exact <- select_by_vernacular(data = bf_data,
                                          names = "pinheiro",
                                          exact = FALSE)
head(pinheiro_not_exact)
#>                         species                              scientificName
#> 26790    Araucaria angustifolia     Araucaria angustifolia (Bertol.) Kuntze
#> 123434     Cryptomeria japonica Cryptomeria japonica (Thunb. ex L.f.) D.Don
#> 123566       Taxodium distichum               Taxodium distichum (L.) Rich.
#> 133878  Cunninghamia lanceolata       Cunninghamia lanceolata (Lamb.) Hook.
#> 134008     Araucaria columnaris     Araucaria columnaris (J.R.Forst.) Hook.
#> 134501 Chamaecyparis lawsoniana   Chamaecyparis lawsoniana (A.Murray) Parl.
#>        acceptedName kingdom       Group Subgroup        family         genus
#> 26790          <NA> Plantae Gymnosperms     <NA> Araucariaceae     Araucaria
#> 123434         <NA> Plantae Gymnosperms     <NA>  Cupressaceae   Cryptomeria
#> 123566         <NA> Plantae Gymnosperms     <NA>  Cupressaceae      Taxodium
#> 133878         <NA> Plantae Gymnosperms     <NA>  Cupressaceae  Cunninghamia
#> 134008         <NA> Plantae Gymnosperms     <NA> Araucariaceae     Araucaria
#> 134501         <NA> Plantae Gymnosperms     <NA>  Cupressaceae Chamaecyparis
#>        lifeForm     habitat                 Biome                     States
#> 26790      Tree Terrestrial Atlantic_Forest;Pampa          MG;PR;RJ;RS;SC;SP
#> 123434     Tree Terrestrial                       DF;ES;MG;PA;PR;RJ;RS;SC;SP
#> 123566     Tree Terrestrial                                MG;PR;RJ;RS;SC;SP
#> 133878     Tree Terrestrial                                MG;PR;RJ;RS;SC;SP
#> 134008     Tree Terrestrial                          MG;PB;PE;PR;RS;SC;SE;SP
#> 134501     Tree Terrestrial                                      PR;RJ;RS;SP
#>                                                                            vegetationType
#> 26790  High_Altitude_Grassland;Mixed_Ombrophyllous_Forest;Seasonally_Semideciduous_Forest
#> 123434                                                                     Anthropic_Area
#> 123566                                                                     Anthropic_Area
#> 133878                                                                     Anthropic_Area
#> 134008                                                                     Anthropic_Area
#> 134501                                                                     Anthropic_Area
#>            Origin    Endemism taxonomicStatus nomenclaturalStatus
#> 26790      Native Non-endemic        Accepted             Correct
#> 123434 Cultivated Non-endemic        Accepted             Correct
#> 123566 Cultivated Non-endemic        Accepted             Correct
#> 133878 Cultivated Non-endemic        Accepted             Correct
#> 134008 Cultivated Non-endemic        Accepted             Correct
#> 134501 Cultivated Non-endemic        Accepted                <NA>
#>                                                                                                                                                                                                vernacularName
#> 26790                                                                                                                               araucaria, pinheiro-do-parana, curi, pinheiro-brasileiro, pinho-do-parana
#> 123434                                                                                                cedro-japones, pinheiro-do-japao, araucaria-do-japao, arvore-de-natal, japanese-cedar, cedrinho-japones
#> 123566 cipreste-calvo, cipreste-deciduo, cipreste-do-brejo, pinheiro-do-brejo, cipreste-vermelho, swamp-cypress, louisiana-cypress, marsh-cypress, cispreste-da-luisiana, bald-cypress, cipreste-dos-pantanos
#> 133878                                                                                                             cuningamia, cunninghamia, pinhao-chines, pinheiro-alemao, pinheiro-chines, arvore-do-natal
#> 134008                                                                                                   pinheiro, pinheiro-de-natal, pinheiro-alemao, pinheiro-de-cook, araucaria-colunar, araucaria-excelsa
#> 134501                                     cedro-de-oregon, cerdo-branco, tuia, falso-cipreste, pinheiro-prateado, cipreste-de-lawson, cipreste-nevado, cipreste, alumi, cipreste-alumi, cipreste-escaravelho
#>        taxonRank
#> 26790    Species
#> 123434   Species
#> 123566   Species
#> 133878   Species
#> 134008   Species
#> 134501   Species