Extract a subset of species from Brazilian Flora 2020 database

Returns a data.frame with a subset of species from Brazilian Flora 2020 database

Usage

subset_species(data = NULL, species = NULL,
                      include_subspecies = FALSE,
                      include_variety = FALSE,
                      Kingdom = "Plantae")

Arguments

data

(data.frame) the data.frame imported with the load_florabr function.

species

(character) names of the species to be extracted from Brazilian Flora database.

include_subspecies

(logical) include subspecies? Default = FALSE

include_variety

(logical) include varieties of the species? Default = FALSE

Kingdom

(character) The Kingdom for filtering the dataset. It can be "Plantae" or "Fungi". Default = "Plantae". To include both, use c("Plantae", "Fungi")

Value

A data.frame with the selected species.

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Brazilian Flora data
#Species to extract from database
spp <- c("Araucaria angustifolia", "Adesmia paranensis")
spp_bf <- subset_species(data = bf_data, species = spp,
                      include_subspecies = FALSE,
                      include_variety = FALSE)
spp_bf
#>                      species                          scientificName
#> 26790 Araucaria angustifolia Araucaria angustifolia (Bertol.) Kuntze
#> 82029     Adesmia paranensis              Adesmia paranensis Burkart
#>       acceptedName kingdom       Group Subgroup        family     genus
#> 26790         <NA> Plantae Gymnosperms     <NA> Araucariaceae Araucaria
#> 82029         <NA> Plantae Angiosperms     <NA>      Fabaceae   Adesmia
#>       lifeForm     habitat                 Biome            States
#> 26790     Tree Terrestrial Atlantic_Forest;Pampa MG;PR;RJ;RS;SC;SP
#> 82029 Subshrub Terrestrial       Atlantic_Forest             PR;SC
#>                                                                           vegetationType
#> 26790 High_Altitude_Grassland;Mixed_Ombrophyllous_Forest;Seasonally_Semideciduous_Forest
#> 82029                                                            High_Altitude_Grassland
#>       Origin    Endemism taxonomicStatus nomenclaturalStatus
#> 26790 Native Non-endemic        Accepted             Correct
#> 82029 Native     Endemic        Accepted             Correct
#>                                                                  vernacularName
#> 26790 araucaria, pinheiro-do-parana, curi, pinheiro-brasileiro, pinho-do-parana
#> 82029                                                                      <NA>
#>       taxonRank
#> 26790   Species
#> 82029   Species