Get available attributes to filter species

This function displays all the options available to filter species by its characteristics

get_attributes(data, attribute, Kingdom = "Plantae")

Arguments

data

(data.frame) a data.frame imported with the load_florabr function or a data.frame generated with the select_species function.

attribute

(character) the type of characteristic. See detail to see the options.

Kingdom

(character) the kingdom to which the species belong. It can be "Plantae" or "Fungi". Default = "Plantae".

Value

a data.frame with two columns. The first column provides the available options in English. Use this options in the select_species function. The second columns provides the options in Portuguese.

Details

The attribute argument accepts the following options: Group, SubGroup, family, lifeForm, habitat, vegetationType, Origin, Endemism, Biome, States, taxonomicStatus or nomenclaturalStatus". These options represent different characteristics of species that can be used for filtering.

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Brazilian Flora data
# Get available biomes to filter species
get_attributes(data = bf_data, Kingdom = "Plantae", attribute = "Biome")
#>                 Biome                Bioma
#> 1              Amazon             Amazonia
#> 2     Atlantic_Forest       Mata_Atlantica
#> 3            Caatinga             Caatinga
#> 4             Cerrado              Cerrado
#> 5               Pampa                Pampa
#> 6            Pantanal             Pantanal
#> 7 Not_found_in_Brazil Nao_ocorre_no_Brasil
# Get available life forms to filter species
get_attributes(data = bf_data, Kingdom = "Plantae", attribute = "lifeForm")
#>               lifeForm            Forma_de_vida
#> 1      Aquatic-Benthos          Aquatica-Bentos
#> 2      Aquatic-Neuston         Aquatica-Neuston
#> 3     Aquatic-Plankton        Aquatica-Plancton
#> 4                Shrub                  Arbusto
#> 5                 Tree                   Arvore
#> 6               Bamboo                    Bambu
#> 7              Cushion                    Coxim
#> 8             Dendroid                Dendroide
#> 9              Unknown             Desconhecida
#> 10          Dracaenoid               Dracenoide
#> 13                Herb                     Erva
#> 14          Flabellate                Flabelado
#> 15             Foliose                  Folhosa
#> 16 Liana/scandent/vine Liana/voluvel/trepadeira
#> 19           Palm_tree                 Palmeira
#> 21             Pendent                 Pendente
#> 23            Subshrub               Subarbusto
#> 24           Succulent                Suculenta
#> 25            Thallose                   Talosa
#> 26                 Mat                   Tapete
#> 27                Weft                    Trama
#> 28                Tuft                     Tufo
# Get available states to filter species
get_attributes(data = bf_data, Kingdom = "Plantae", attribute = "States")
#>    Sigla_Acronym           Nome_Name
#> 1             AC                Acre
#> 2             AM            Amazonas
#> 3             PA                Para
#> 4             BA               Bahia
#> 5             ES      Espirito Santo
#> 6             RJ      Rio de Janeiro
#> 7             MT         Mato Grosso
#> 8             MA            Maranhao
#> 9             CE               Ceara
#> 10            MG        Minas Gerais
#> 11            PE          Pernambuco
#> 12            PI               Piaui
#> 13            SE             Sergipe
#> 14            MS  Mato Grosso do Sul
#> 15            PR              Parana
#> 16            RS   Rio Grande do Sul
#> 17            SC      Santa Catarina
#> 18            SP           Sao Paulo
#> 19            AL             Alagoas
#> 20            DF    Distrito Federal
#> 21            GO               Goias
#> 22            RO            Rondania
#> 23            AP               Amapa
#> 24            PB             Paraiba
#> 25            RN Rio Grande do Norte
#> 26            RR             Roraima
#> 27            TO           Tocantins