Retrieve synonyms for species

Retrieve synonyms for species

get_synonym(data, species,
                   include_subspecies = TRUE, include_variety = TRUE)

Arguments

data

(data.frame) the data.frame imported with the load_florabr function

species

(character) names of the species

include_subspecies

(logical) include subspecies that are synonyms of the species? Default = TRUE

include_variety

(logical) include varieties that are synonyms of the species? Default = TRUE

Value

A data.frame containing unique synonyms of the specified species along with relevant information on taxonomic and nomenclatural statuses.

References

Flora e Funga do Brasil. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Flora e Funga do Brasil data
#Species to extract synonyms
spp <- c("Araucaria angustifolia", "Adesmia paranensis")
spp_synonyms <- get_synonym(data = bf_data, species = spp,
                            include_subspecies = TRUE,
                            include_variety = TRUE)
spp_synonyms
#>                  acceptedName                                    synonym
#> 35312  Araucaria angustifolia           Araucaria angustifolia var. alba
#> 35313  Araucaria angustifolia         Araucaria angustifolia var. caiova
#> 35314  Araucaria angustifolia         Araucaria angustifolia var. caiuva
#> 35315  Araucaria angustifolia      Araucaria angustifolia var. dependens
#> 35316  Araucaria angustifolia    Araucaria angustifolia var. indehiscens
#> 35317  Araucaria angustifolia        Araucaria angustifolia var. monoica
#> 35318  Araucaria angustifolia          Araucaria angustifolia var. nigra
#> 35319  Araucaria angustifolia Araucaria angustifolia var. sancti-josephi
#> 35320  Araucaria angustifolia       Araucaria angustifolia var. semialba
#> 35321  Araucaria angustifolia        Araucaria angustifolia var. stricta
#> 35322  Araucaria angustifolia        Araucaria angustifolia var. vinacea
#> 35323  Araucaria angustifolia                       Araucaria brasiliana
#> 35324  Araucaria angustifolia          Araucaria brasiliana var. elegans
#> 35325  Araucaria angustifolia                     Araucaria brasiliensis
#> 35326  Araucaria angustifolia        Araucaria brasiliensis var. saviana
#> 35327  Araucaria angustifolia                           Araucaria dioica
#> 35328  Araucaria angustifolia                          Araucaria elegans
#> 35329  Araucaria angustifolia                       Araucaria ridolfiana
#> 35330  Araucaria angustifolia                          Araucaria saviana
#> 35332  Araucaria angustifolia                      Columbea angustifolia
#> 35333  Araucaria angustifolia                        Columbea brasiliana
#> 60644  Araucaria angustifolia                               Pinus dioica
#> 141010 Araucaria angustifolia       Araucaria brasiliana var. ridolfiana
#> 141020 Araucaria angustifolia                         Araucaria bibbiani
#> 141021 Araucaria angustifolia                       Araucaria lindleyana
#> 141041 Araucaria angustifolia                       Araucaria brasiliana
#> 85308      Adesmia paranensis         Adesmia psoraleoides var. glabrata
#>        taxonomicStatus      nomenclaturalStatus
#> 35312          Synonym                     <NA>
#> 35313          Synonym                     <NA>
#> 35314          Synonym                     <NA>
#> 35315          Synonym                     <NA>
#> 35316          Synonym                     <NA>
#> 35317          Synonym                     <NA>
#> 35318          Synonym                     <NA>
#> 35319          Synonym                     <NA>
#> 35320          Synonym                     <NA>
#> 35321          Synonym                     <NA>
#> 35322          Synonym                     <NA>
#> 35323          Synonym                     <NA>
#> 35324          Synonym                     <NA>
#> 35325          Synonym                     <NA>
#> 35326          Synonym                     <NA>
#> 35327          Synonym                     <NA>
#> 35328          Synonym                     <NA>
#> 35329          Synonym                     <NA>
#> 35330          Synonym                     <NA>
#> 35332          Synonym Legitimate_but_incorrect
#> 35333          Synonym Legitimate_but_incorrect
#> 60644          Synonym                     <NA>
#> 141010         Synonym                     <NA>
#> 141020         Synonym                     <NA>
#> 141021         Synonym                     <NA>
#> 141041         Synonym             Illegitimate
#> 85308          Synonym Legitimate_but_incorrect