Search for taxa using vernacular names

Search for taxa using vernacular names

select_by_vernacular(data, names, exact = FALSE)

Arguments

data

(data.frame) the data.frame imported with the load_florabr function or generated with the function select_species.

names

(character) vernacular name ("Nome comum") of the species to be searched

exact

(logic) if TRUE, the function will search only for exact matches. For example, if names = "pinheiro" and exact = TRUE, the function will return only the species popularly known as "pinheiro". On the other hand, if names = "pinheiro" and exact = FALSE, the function will return other results as "pinheiro-do-parana". Default = FALSE

Value

a data.frame with the species with vernacular names that match the input names

References

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Brazilian Flora 2020. Jardim Botânico do Rio de Janeiro. Available at: http://floradobrasil.jbrj.gov.br/

Examples

data("bf_data") #Load Brazilian Flora data
#Search for species whose vernacular name is 'pinheiro'
pinheiro_exact <- select_by_vernacular(data = bf_data,
                                       names = "pinheiro",
                                       exact = TRUE)
pinheiro_exact
#>                     species
#> 134218 Araucaria columnaris
#> 140840    Pinus canariensis
#> 140842      Pinus banksiana
#> 140856      Pinus palustris
#> 140864          Pinus nigra
#> 140890       Pinus echinata
#> 140902     Pinus densiflora
#> 140986    Pinus tecunumanii
#> 140991     Pinus thunbergii
#>                                               scientificName acceptedName
#> 134218               Araucaria columnaris (J.R.Forst.) Hook.         <NA>
#> 140840                               Pinus canariensis C.Sm.         <NA>
#> 140842                                 Pinus banksiana Lamb.         <NA>
#> 140856                                 Pinus palustris Mill.         <NA>
#> 140864                                Pinus nigra J.F.Arnold         <NA>
#> 140890                                  Pinus echinata Mill.         <NA>
#> 140902                      Pinus densiflora Siebold & Zucc.         <NA>
#> 140986 Pinus tecunumanii F.Schwerdtf. ex Eguiluz & J.P.Perry         <NA>
#> 140991                                Pinus thunbergii Parl.         <NA>
#>        kingdom       group subgroup       phylum     class   order
#> 134218 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140840 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140842 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140856 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140864 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140890 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140902 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140986 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 140991 Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#>               family     genus lifeForm     habitat   biome
#> 134218 Araucariaceae Araucaria     Tree Terrestrial Unknown
#> 140840      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140842      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140856      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140864      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140890      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140902      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140986      Pinaceae     Pinus     Tree Terrestrial Unknown
#> 140991      Pinaceae     Pinus     Tree Terrestrial Unknown
#>                         states     vegetation      origin    endemism
#> 134218 MG;PB;PE;PR;RS;SC;SE;SP Anthropic_Area  Cultivated Non-endemic
#> 140840             MG;RJ;RS;SP Anthropic_Area  Cultivated Non-endemic
#> 140842                      RS Anthropic_Area  Cultivated Non-endemic
#> 140856             PR;RS;SC;SP Anthropic_Area  Cultivated Non-endemic
#> 140864                   RS;SP Anthropic_Area  Cultivated Non-endemic
#> 140890                PR;RS;SP Anthropic_Area  Cultivated Non-endemic
#> 140902                      MG Anthropic_Area  Cultivated Non-endemic
#> 140986                      SP Anthropic_Area Naturalized Non-endemic
#> 140991                   PR;SP Anthropic_Area  Cultivated Non-endemic
#>        taxonomicStatus nomenclaturalStatus
#> 134218        Accepted             Correct
#> 140840        Accepted                <NA>
#> 140842        Accepted                <NA>
#> 140856        Accepted                <NA>
#> 140864        Accepted                <NA>
#> 140890        Accepted                <NA>
#> 140902        Accepted                <NA>
#> 140986        Accepted                <NA>
#> 140991        Accepted                <NA>
#>                                                                                                                                                     vernacularName
#> 134218                                                        pinheiro, pinheiro-de-natal, pinheiro-alemao, pinheiro-de-cook, araucaria-colunar, araucaria-excelsa
#> 140840                                                                                                                             pinheiro-das-canarias, pinheiro
#> 140842                                                                                                                                  pinus, pinheiro, jack-pine
#> 140856                          pinheiro, pinheiro-de-folha-longa, pinheiro-da-florida, pinheiro-amarelo, pinheiro-breu, pinheiro-da-georgia, pinheiro-terebintina
#> 140864                                                                                                                                     pinheiro, austrian-pine
#> 140890 pinheiro-de-folha-curta, pinheiro-da-carolina-do-norte, pinheiro-pobre, pinheiro-amarelo-de-folha-curta, pinheiro-amarelo, pinus, pinheiro, short-leaf-pine
#> 140902                                                                                                                                 pinheiro, japanese-red-pine
#> 140986                                                                                                                                             pinus, pinheiro
#> 140991                                                                                                                                             pinus, pinheiro
#>        taxonRank     id
#> 134218   Species 609385
#> 140840   Species 616545
#> 140842   Species 616547
#> 140856   Species 616568
#> 140864   Species 616576
#> 140890   Species 616608
#> 140902   Species 616620
#> 140986   Species 616713
#> 140991   Species 616718
#Search for species whose vernacular name is 'pinheiro', allowing non-exact
#matches
pinheiro_not_exact <- select_by_vernacular(data = bf_data,
                                          names = "pinheiro",
                                          exact = FALSE)
head(pinheiro_not_exact)
#>                         species                              scientificName
#> 26797    Araucaria angustifolia     Araucaria angustifolia (Bertol.) Kuntze
#> 123595     Cryptomeria japonica Cryptomeria japonica (Thunb. ex L.f.) D.Don
#> 123731       Taxodium distichum               Taxodium distichum (L.) Rich.
#> 134088  Cunninghamia lanceolata       Cunninghamia lanceolata (Lamb.) Hook.
#> 134218     Araucaria columnaris     Araucaria columnaris (J.R.Forst.) Hook.
#> 134711 Chamaecyparis lawsoniana   Chamaecyparis lawsoniana (A.Murray) Parl.
#>        acceptedName kingdom       group subgroup       phylum     class   order
#> 26797          <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 123595         <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 123731         <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 134088         <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 134218         <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#> 134711         <NA> Plantae Gymnosperms     <NA> Tracheophyta Pinopsida Pinales
#>               family         genus lifeForm     habitat                 biome
#> 26797  Araucariaceae     Araucaria     Tree Terrestrial Atlantic_Forest;Pampa
#> 123595  Cupressaceae   Cryptomeria     Tree Terrestrial               Unknown
#> 123731  Cupressaceae      Taxodium     Tree Terrestrial               Unknown
#> 134088  Cupressaceae  Cunninghamia     Tree Terrestrial               Unknown
#> 134218 Araucariaceae     Araucaria     Tree Terrestrial               Unknown
#> 134711  Cupressaceae Chamaecyparis     Tree Terrestrial               Unknown
#>                            states
#> 26797           MG;PR;RJ;RS;SC;SP
#> 123595 DF;ES;MG;PA;PR;RJ;RS;SC;SP
#> 123731          MG;PR;RJ;RS;SC;SP
#> 134088          MG;PR;RJ;RS;SC;SP
#> 134218    MG;PB;PE;PR;RS;SC;SE;SP
#> 134711                PR;RJ;RS;SP
#>                                                                                vegetation
#> 26797  High_Altitude_Grassland;Mixed_Ombrophyllous_Forest;Seasonally_Semideciduous_Forest
#> 123595                                                                     Anthropic_Area
#> 123731                                                                     Anthropic_Area
#> 134088                                                                     Anthropic_Area
#> 134218                                                                     Anthropic_Area
#> 134711                                                                     Anthropic_Area
#>            origin    endemism taxonomicStatus nomenclaturalStatus
#> 26797      Native Non-endemic        Accepted             Correct
#> 123595 Cultivated Non-endemic        Accepted             Correct
#> 123731 Cultivated Non-endemic        Accepted             Correct
#> 134088 Cultivated Non-endemic        Accepted             Correct
#> 134218 Cultivated Non-endemic        Accepted             Correct
#> 134711 Cultivated Non-endemic        Accepted                <NA>
#>                                                                                                                                                                                                vernacularName
#> 26797                                                                                                                               araucaria, pinheiro-do-parana, curi, pinheiro-brasileiro, pinho-do-parana
#> 123595                                                             pinheiro-vermelho, cedro-do-japao, cedro-japones, pinheiro-do-japao, araucaria-do-japao, arvore-de-natal, japanese-cedar, cedrinho-japones
#> 123731 cipreste-calvo, cipreste-deciduo, cipreste-do-brejo, pinheiro-do-brejo, cipreste-vermelho, swamp-cypress, louisiana-cypress, marsh-cypress, cispreste-da-luisiana, bald-cypress, cipreste-dos-pantanos
#> 134088                                                                                                             cuningamia, cunninghamia, pinhao-chines, pinheiro-alemao, pinheiro-chines, arvore-do-natal
#> 134218                                                                                                   pinheiro, pinheiro-de-natal, pinheiro-alemao, pinheiro-de-cook, araucaria-colunar, araucaria-excelsa
#> 134711                 cedro-de-oregon, cerdo-branco, cipreste-de-lawson, tuia, falso-cipreste, pinheiro-prateado, cipreste-de-lawson, cipreste-nevado, cipreste, alumi, cipreste-alumi, cipreste-escaravelho
#>        taxonRank     id
#> 26797    Species  33971
#> 123595   Species 596703
#> 123731   Species 596865
#> 134088   Species 609253
#> 134218   Species 609385
#> 134711   Species 609897